Extract Loading and Error panel states to components folder
